file is modified

+++ assignment1_4.cpp
@@ -1,18 +1,18 @@
 #include <iostream>
 using namespace std;
-inline float triangleArea(float base, float height) 
+inline double triangleArea(double base, double height) 
 {
     return 0.5 * base * height;
 }
 
 int main() 
 {
-    float base, height;
+    double base, height;
     cout << "Enter the base of the triangle: ";
     cin >> base;
     cout << "Enter the height of the triangle: ";
     cin >> height;
-    float area = triangleArea(base, height);
+    double area = triangleArea(base, height);
     cout << "The area of the triangle is: " << area << std::endl;
     return 0;
 }
